Finished implementing remove node script for munin

+++ plugins.d/munin/removenode.d/helpscript/removenode.py
@@ -1,15 +1,4 @@
 #/usr/bin/env python
-"""
-xxxx
-[para102]
-        address 10.46.0.102
-        use_node_name yes
-
-
-
-"""
-
-
 import sys
 
 ip = sys.argv[1]
@@ -17,13 +6,15 @@ ip = sys.argv[1]
 with open('/etc/munin/munin.conf', 'r+') as f:
   data=f.read()
   f.seek(0)
-  data = re.sub(
-            re.compile(r"(\[[A-Za-z0-9]+\])"
-                       ""
-                       "([\t ]+)address "+ip
-                       "", re.MULTILINE), replaceWith, data
-            )
-  #print data
-  f.write(data)
+  #find the node to be removed
+  regex = re.compile(r"^(\[(.+)\])((\n\t(.+)(\s)(.+))*)(\n\t(address)(\s))"+ip+"((\n\t(.+)(\s)(.+))*)", re.MULTILINE)
+  #replace node configuration with the empty string
+  newconf = re.sub(regex, "", data)
+  
+  if newconf == data:
+    print "no node could be found"
+    sys.exit()
+  
+  f.write(newconf)
   f.truncate()
   f.close()
